requirements.txt
redis
python-jose[cryptography]
passlib[bcrypt]
pytest-asyncio==0.23.2
pytest==7.4.3
marshmallow==3.20.1
nltk==3.8.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastapi==0.119.0
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.7.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
statsmodels==0.14.5
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
fastapi
uvicorn[standard]
streamlit
pandas
python-dotenv
groq
statsmodels
matplotlib
requests
python-multipart
lightgbm
scikit-learn
crewai==0.86.0
crewai-tools==0.17.0
langchain==0.3.18
langchain-groq==0.2.8